sdm660-common: props: Disable vsync for CPU rendered Apps
* Choreographer VSync has some performance impact for CPU rendered apps
as the app has to wait for VSync to start a new frame draw. When the
frame draw time exceeds one VSync, this has a direct impact on the FPS.
debug.cpurend.vsync property controls the VSync for CPU rendered apps.

Disable VSync for CPU rendered app, by setting this property to false.

Signed-off-by: Vitorgl2003 <vitorgl.2003@gmail.com>
Signed-off-by: xenxynon <flynryder427@gmail.com>
Signed-off-by: ImPrashantt <prashant33968@gmail.com>